fixed function showScreen; removed actions with DOM from templates, moved it in bind method of view
