+++ js/templates/loss.js
@@ -1,16 +1,13 @@
 import LossView from '../view/loss-view';
-import {showScreen, container} from '../functions';
-import gameCurrentState from './game-state';
+import {showScreen, removeGameProgressContainer, removeGameStateContainer} from '../functions';
 
 const moduleLoss = () => {
-  gameCurrentState();
+  removeGameProgressContainer();
+  removeGameStateContainer();
 
   const loss = new LossView();
 
   showScreen(loss.element);
-
-  /** TODO убрать хардкорное удаление блока с игровым прогрессом **/
-  container.nextElementSibling.remove();
 };
 
 export default moduleLoss;
